feat: event preliminary.3 create/get events; eventRequest create/get; before admin event status change endpoints

+++ main-service/src/main/java/ru/yandex/practicum/event/model/Event.java
@@ -60,7 +60,7 @@ public class Event {
     boolean paid;
 
     @Column(name = "participant_limit")
-    int participantLimit;
+    long participantLimit;
 
     @Column(name = "published_on")
     @DateTimeFormat(pattern = "yyyy-MM-dd HH:mm:ss")
