+++ main-service/src/main/java/ru/yandex/practicum/event/model/dto/EventFullInfoDto.java
@@ -22,7 +22,7 @@ public class EventFullInfoDto {
 
     CategoryDto category;
 
-    int confirmedRequests;
+    long confirmedRequests;
 
     @DateTimeFormat(pattern = "yyyy-MM-dd HH:mm:ss")
     @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss")
@@ -41,7 +41,7 @@ public class EventFullInfoDto {
 
     boolean paid;
 
-    int participantLimit;
+    long participantLimit;
 
     @DateTimeFormat(pattern = "yyyy-MM-dd HH:mm:ss")
     @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss")
